// helpers/types.ts
export enum eContractid {
  Example = 'Example',
  LendingPoolAddressesProvider = 'LendingPoolAddressesProvider',
  MintableERC20 = 'MintableERC20',
  MintableDelegationERC20 = 'MintableDelegationERC20',
  LendingPoolAddressesProviderRegistry = 'LendingPoolAddressesProviderRegistry',
  LendingPoolParametersProvider = 'LendingPoolParametersProvider',
  LendingPoolConfigurator = 'LendingPoolConfigurator',
  ValidationLogic = 'ValidationLogic',
  ReserveLogic = 'ReserveLogic',
  GenericLogic = 'GenericLogic',
  LendingPool = 'LendingPool',
  PriceOracle = 'PriceOracle',
  Proxy = 'Proxy',
  MockAggregator = 'MockAggregator',
  MockAggregatorDIA = 'MockAggregatorDIA',
  LendingRateOracle = 'LendingRateOracle',
  SiO2Oracle = 'SiO2Oracle',
  SiO2FallbackOracle = 'SiO2FallbackOracle',
  DefaultReserveInterestRateStrategy = 'DefaultReserveInterestRateStrategy',
  LendingPoolCollateralManager = 'LendingPoolCollateralManager',
  InitializableAdminUpgradeabilityProxy = 'InitializableAdminUpgradeabilityProxy',
  MockFlashLoanReceiver = 'MockFlashLoanReceiver',
  WalletBalanceProvider = 'WalletBalanceProvider',
  SToken = 'SToken',
  MockSToken = 'MockSToken',
  DelegationAwareSToken = 'DelegationAwareSToken',
  MockStableDebtToken = 'MockStableDebtToken',
  MockVariableDebtToken = 'MockVariableDebtToken',
  SiO2ProtocolDataProvider = 'SiO2ProtocolDataProvider',
  IERC20Detailed = 'IERC20Detailed',
  StableDebtToken = 'StableDebtToken',
  VariableDebtToken = 'VariableDebtToken',
  FeeProvider = 'FeeProvider',
  TokenDistributor = 'TokenDistributor',
  StableAndVariableTokensHelper = 'StableAndVariableTokensHelper',
  STokensAndRatesHelper = 'STokensAndRatesHelper',
  UiPoolDataProviderV2 = 'UiPoolDataProviderV2',
  UiPoolDataProviderV2Proxy = 'UiPoolDataProviderV2Proxy',
  WETHGateway = 'WETHGateway',
  WETH = 'WETH',
  WASTRMocked = 'WASTRMocked',
  SelfdestructTransferMock = 'SelfdestructTransferMock',
  LendingPoolImpl = 'LendingPoolImpl',
  LendingPoolConfiguratorImpl = 'LendingPoolConfiguratorImpl',
  LendingPoolCollateralManagerImpl = 'LendingPoolCollateralManagerImpl',
  PriceAggregatorAdapterRedStoneImpl = 'PriceAggregatorAdapterRedStoneImpl',
  RedStoneTransparentUpgradeableProxy = 'RedStoneTransparentUpgradeableProxy',
  UiIncentiveDataProviderV2 = 'UiIncentiveDataProviderV2',
  UiIncentiveDataProviderV2Proxy = 'UiIncentiveDataProviderV2Proxy',
  StakeUIHelper = 'StakeUIHelper',
  StakeUIHelperProxy = 'StakeUIHelperProxy',
  IUniswapV2Router02 = 'IUniswapV2Router02',
  IUniswapV2Factory = 'IUniswapV2Factory',
  FlashLiquidationAdapter = 'FlashLiquidationAdapter',
  MockedSiO2IncentivesController = 'MockedSiO2IncentivesController',

}
// ...
export type tEthereumAddress = string;
// ...
export type iSiO2PoolAssets<T> = Pick<
  iAssetsWithoutUSD<T>,
  'USDC' | 
  // 'WBTC' | 
  'WETH' | 
  // 'WASTR' | 
  // 'BUSD' | 
  // 'BAI' | 
  'USDT' | 
  // 'DAI' | 
  // 'DOT' | 
  // 'BNB' | 
  // 'nASTR' | 
  'TIA'
>;

export type iMultiPoolsAssets<T> = iAssetCommon<T> | iSiO2PoolAssets<T>;

export type iSiO2PooSTokens<T> = Omit<iSiO2PoolAssets<T>, 'ETH'>;

export const TokenContractId = [
  'SIO2',
  'WETH',
  'USDC',
  // 'WBTC',
  'USD',
  // 'WASTR',
  'USDT',
  'TIA'
  // 'BUSD',
  // 'BAI',
  // 'DAI',
  // 'WDOT',
  // 'WBNB',
  // 'nASTR',
] as const;

export interface IReserveParams extends IReserveBorrowParams, IReserveCollateralParams {
  STokenImpl: eContractid;
  reserveFactor: string;
  strategy: IInterestRateStrategyParams;
}
// ...
export interface IMocksConfig {
  AllAssetsInitialPrices: iAssetBase<string>;
}
// ...
export interface IBaseConfiguration {
  MarketId: string;
  STokenNamePrefix: string;
  StableDebtTokenNamePrefix: string;
  VariableDebtTokenNamePrefix: string;
  SymbolPrefix: string;
  ProviderId: number;
  ProtocolGlobalParams: IProtocolGlobalConfig;
  ProviderRegistry: iParamsPerNetwork<tEthereumAddress | undefined>;
  ProviderRegistryOwner: iParamsPerNetwork<tEthereumAddress | undefined>;
  LendingPoolCollateralManager: iParamsPerNetwork<tEthereumAddress>;
  LendingPoolConfigurator: iParamsPerNetwork<tEthereumAddress>;
  LendingPool: iParamsPerNetwork<tEthereumAddress>;
  LendingRateOracleRatesCommon: iMultiPoolsAssets<IMarketRates>;
  LendingRateOracle: iParamsPerNetwork<tEthereumAddress>;
  TokenDistributor: iParamsPerNetwork<tEthereumAddress>;
  SiO2Oracle: iParamsPerNetwork<tEthereumAddress>;
  PriceAggregator: iParamsPerNetwork<tEthereumAddress>;
  FallbackOracle: iParamsPerNetwork<tEthereumAddress>;
  ChainlinkAggregator: iParamsPerNetwork<ITokenAddress>;
  DIAAggregatorAddress: iParamsPerNetwork<tEthereumAddress>;
  PoolAdmin: iParamsPerNetwork<tEthereumAddress | undefined>;
  PoolAdminIndex: number;
  EmergencyAdmin: iParamsPerNetwork<tEthereumAddress | undefined>;
  EmergencyAdminIndex: number;
  STokenDomainSeparator: iParamsPerNetwork<string>;
  WETH: iParamsPerNetwork<tEthereumAddress>;
  WrappedNativeToken: iParamsPerNetwork<tEthereumAddress>;
  WethGateway: iParamsPerNetwork<tEthereumAddress>;
  ReserveFactorTreasuryAddress: iParamsPerNetwork<tEthereumAddress>;
  IncentivesController: iParamsPerNetwork<tEthereumAddress>;
  StableDebtTokenImplementation?: iParamsPerNetwork<tEthereumAddress>;
  VariableDebtTokenImplementation?: iParamsPerNetwork<tEthereumAddress>;
  ReserveAssets: iParamsPerNetwork<SymbolMap<tEthereumAddress>>;
  OracleQuoteCurrency: string;
  OracleQuoteUnit: string;
  StakedSiO2: iParamsPerNetwork<tEthereumAddress>;
  IncentiveRewardStrategy: iParamsPerNetwork<IIncentiveRewardStrategy>;
}

export interface ICommonConfiguration extends IBaseConfiguration {
  ReservesConfig: iMultiPoolsAssets<IReserveParams>;
  Mocks: IMocksConfig;
}

export interface ISiO2Configuration extends ICommonConfiguration {
  ReservesConfig: iSiO2PoolAssets<IReserveParams>;
}

export interface ITokenAddress {
  [token: string]: tEthereumAddress;
}

export type PoolConfiguration = ICommonConfiguration | ISiO2Configuration;